Closes #29 Pillboxes fire faster as they become more damaged
Here's the theory behind it: Pillboxes run on an energy source, which powers aiming, firing, replenishing ammo, and repairing. As a pillbox becomes more damaged, it diverts energy into the three offensive functions, at the cost of reducing its repair rate. While all energy has been diverted to its weapon systems, it is unable to repair.

In game, this means that as a pillbox becomes more damaged, it fires faster, but repairs more slowly. At its angriest, it will fire very quickly, but won't repair until it has no targets in range. This differs from the way the original game worked, but may be a fun change.